ci: avoid running git diff after patching
Drop `git diff` command so it is easier to run CI locally if git checkout is a
worktree. Currently it fails because the directory is not recognized as a git
repository.

The `git diff` command was added recently in #28359 commit
b6a1892bc1c1f801200dcb5c8faa7926d6c0c0f1 and can be avoided just by teeing the
patch to stdout