tracing: cast block_connected duration to nanoseconds
When the tracepoint was introduced in 8b5824cd7479655a8b4450d1da225674a6e7e42b,
the connect_block duration was passed in microseconds `µs`.
By starting to use steady clock in ffa9b060a0f66144b67370d91c7e92bdd78174ac
this changed to nanoseconds `ns`. As the test only checked if the
duration value is `> 0` as a plausibility check, this went unnoticed.

I detected this when setting up monitoring for block validation time
as part of the Great Consensus Cleanup Revival discussion.

This change casts the duration explicitly to nanoseconds (as it has been
nanoseconds for the last three releases; switching back now would 'break'
the broken API again; there don't seem to be many users affected), updates
the documentation and adds a check for an upper bound to the tracepoint
interface tests. The upper bound is quite lax as mining the block takes
much longer than connecting the empty test block. It's however able to
detect incorrect duration units passed.