Move BlockMerkleBranch back to merkle.{h,cpp}
The Mining interface uses this function in the next commit
to calculate the coinbase merkle path. Stratum v2 uses
this to send a compact work template.

This partially undoes the change in 896edc87bd51db31f19070b9079cd65fa523bac0,
but is not a revert, because the implementation changed in the meantime.

This commit also documents the function.