test: Remove intermittent and presumed fixed tsan race suppressions
LoadWallet was added in commit 88d725280514e823c45dff9d1d96336c28248784, which
points to a traceback with BerkeleyBatch in it. Now that BDB is removed,
this can be removed as well.

The race in DatabaseBatch was added in commit
b16c1e6b843c41e9ad689a43b40ffb67f0b1eaab, which does not point to a
traceback. This was likely also fixed with the BDB removal.

If not, the suppressions should be added back, mentioning that they are
intermittent and including a traceback and possibly steps to reproduce.