test: Remove intermittent and presumed fixed tsan race suppressions
LoadWallet was added in commit 05bd161b4c89c1ce7e9eee51cd4ec8877839bcf4, which
points to a traceback with BerkeleyBatch in it. Now that BDB is removed,
this can be removed as well.

The race in DatabaseBatch was added in commit
e5740214cee6f2a727f864d54c602decaeb784bf, which does not point to a
traceback. This was likely also fixed with the BDB removal.

If not, the suppressions should be added back, mentioning that they are
intermittent and including a traceback and possibly steps to reproduce.