test: Remove intermittent and presumed fixed tsan race suppressions
LoadWallet was added in commit 5beaa34b1c99ce7c544fdc4f05ad429885cb48ab, which
points to a traceback with BerkeleyBatch in it. Now that BDB is removed,
this can be removed as well.

The race in DatabaseBatch was added in commit
72799c894132bf894836a59b905a2a81f22a1f6b, which does not point to a
traceback. This was likely also fixed with the BDB removal.

If not, the suppressions should be added back, mentioning that they are
intermittent and including a traceback and possibly steps to reproduce.